requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
geopy==2.4.1
overpy==0.7
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pytz==2023.3.post1
python-json-logger==2.0.7
python-decouple==3.8
blinker==1.7.0